+++ src/main/resources/META-INF/dubbo/com.alibaba.dubbo.registry.RegistryFactory
@@ -1,1 +1,1 @@
-nacos=com.alibaba.dubbo.registry.nacos.NacosRegistryFactory
+nacos=org.apache.dubbo.registry.nacos.NacosRegistryFactory
