+++ src/test/java/org/apache/dubbo/demo/consumer/DemoServiceConsumerXmlBootstrap.java
@@ -14,28 +14,29 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.alibaba.dubbo.demo.provider;
+package org.apache.dubbo.demo.consumer;
 
-import com.alibaba.dubbo.config.spring.context.annotation.EnableDubbo;
-import com.alibaba.dubbo.demo.service.DemoService;
+import org.apache.dubbo.demo.service.DemoService;
 
-import org.springframework.context.annotation.AnnotationConfigApplicationContext;
-import org.springframework.context.annotation.PropertySource;
+import org.springframework.context.support.ClassPathXmlApplicationContext;
 
 import java.io.IOException;
 
 /**
- * {@link DemoService} provider demo
+ * {@link DemoService} consumer demo XML bootstrap
  */
-@EnableDubbo(scanBasePackages = "com.alibaba.dubbo.demo.service")
-@PropertySource(value = "classpath:/provider-config.properties")
-public class DemoServiceProviderBootstrap {
+public class DemoServiceConsumerXmlBootstrap {
 
     public static void main(String[] args) throws IOException {
-        AnnotationConfigApplicationContext context = new AnnotationConfigApplicationContext();
-        context.register(DemoServiceProviderBootstrap.class);
+        ClassPathXmlApplicationContext context = new ClassPathXmlApplicationContext();
+        context.setConfigLocation("/META-INF/spring/dubbo-consumer-context.xml");
         context.refresh();
-        System.out.println("DemoService provider is starting...");
+        System.out.println("DemoService consumer (XML) is starting...");
+        DemoService demoService = context.getBean("demoService", DemoService.class);
+        for (int i = 0; i < 10; i++) {
+            System.out.println(demoService.sayName("小马哥（mercyblitz）"));
+        }
         System.in.read();
+        context.close();
     }
 }
