+++ src/test/java/org/apache/dubbo/registry/nacos/NacosRegistryFactoryTest.java
@@ -14,12 +14,12 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.alibaba.dubbo.registry.nacos;
+package org.apache.dubbo.registry.nacos;
 
 /**
  * {@link NacosRegistryFactory} Test
  *
- * @since 2.6.5
+ * @since 2.7.0
  */
 public class NacosRegistryFactoryTest {
 }
